requirements.txt
httpx>=0.25.0
aiogram>=3.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

